Refactor import paths for cache service and update notification service interface
