+++ lib/config/navigation/app_navigation.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:better_one/config/navigation/routes_enum.dart';
 import 'package:better_one/core/constants/notification_constants.dart';
 import 'package:better_one/core/in_memory/in_memory.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/dependency_injection.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/core/utils/methods/methods.dart';
