+++ lib/core/utils/dependency_locator/dependency_injection.dart
@@ -1,5 +1,5 @@
-import 'package:better_one/core/utils/background_service/foreground_task_notification_service.dart';
-import 'package:better_one/core/utils/background_service/notification_background_service_interface.dart';
+import 'package:better_one/core/utils/service/notification_service/flutter_foreground_task_notification_service.dart';
+import 'package:better_one/core/utils/service/notification_service/notification_background_service_interface.dart';
 import 'package:better_one/core/utils/remote_service/api_consumer/api_consumer.dart';
 import 'package:better_one/core/utils/remote_service/api_consumer/dio_consumer.dart';
 import 'package:better_one/data_source/notification_data_source/firebase_notification_source.dart';
@@ -23,7 +23,7 @@ import 'package:get_it/get_it.dart';
 
 import '../../../repositories/setting_repo/setting_repo.dart';
 import '../../../repositories/setting_repo/settings_repo_interface.dart';
-import '../cache_service/cache_service.dart';
+import '../service/cache_service/cache_service.dart';
 import '../notification_service/flutter_local_notification.dart';
 
 GetIt _getIt = GetIt.instance;
@@ -45,7 +45,7 @@ Future<void> initDependency() async {
       .registerSingleton<NotificationBackgroundService>(
         ForegroundTaskNotificationService(),
       )
-      .initializeService();
+      .initializeAndStartService();
 }
 
 void userDependency() {
