+++ lib/core/utils/service/cache_service/cache_service.dart
@@ -1,6 +1,6 @@
 export 'hive_method/hive_init.dart';
 export 'cach_interface/locale_task_interface.dart';
-export '../../../data_source/task_data_source/hive_locale_task.dart';
-export '../../../data_source/settings_data_source/hive_locale_settings.dart';
+export '../../../../data_source/task_data_source/hive_locale_task.dart';
+export '../../../../data_source/settings_data_source/hive_locale_settings.dart';
 export 'cach_interface/locale_user_info.dart';
 export 'hive_method/hive_locale_user_info.dart';
