+++ lib/data_source/user_data_source/firebase_remote_user_source.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:better_one/core/constants/firebase_constants.dart';
 import 'package:better_one/core/errors/failure.dart';
 import 'package:better_one/core/result_handler/result_handler.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/data_source/user_data_source/remote_user_source.dart';
 import 'package:better_one/model/comment_model/comment_model.dart';
