+++ lib/data_source/user_data_source/hive_locale_user_operations.dart
@@ -5,7 +5,7 @@ import 'package:better_one/core/constants/constants.dart';
 import 'package:better_one/core/enum/task_status.dart';
 import 'package:better_one/core/errors/failure.dart';
 import 'package:better_one/core/result_handler/result_handler.dart';
-import 'package:better_one/core/utils/cache_service/cache_service.dart';
+import 'package:better_one/core/utils/service/cache_service/cache_service.dart';
 import 'package:better_one/core/utils/dependency_locator/dependency_injection.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/data_source/user_data_source/locale_user_source.dart';
