+++ lib/repositories/user_repo/user_repo_impl.dart
@@ -1,7 +1,7 @@
 import 'package:better_one/core/enum/task_status.dart';
 import 'package:better_one/core/errors/failure.dart';
 import 'package:better_one/core/result_handler/result_handler.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/data_source/user_data_source/locale_user_source.dart';
 import 'package:better_one/data_source/user_data_source/remote_user_source.dart';
