+++ lib/view/pages/auth/login_page.dart
@@ -1,8 +1,8 @@
 import 'package:better_one/config/navigation/routes_enum.dart';
 import 'package:better_one/core/constants/constants.dart';
-import 'package:better_one/core/utils/background_service/notification_background_service_interface.dart';
-import 'package:better_one/core/utils/background_service/tasks_background_service.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/notification_service/notification_background_service_interface.dart';
+import 'package:better_one/core/utils/service/task_background_service/tasks_background_service.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/core/utils/methods/methods.dart';
 import 'package:better_one/core/utils/shared_widgets/back_button_l10n.dart';
@@ -114,7 +114,7 @@ class _LoginState extends State<LogIn> {
                       TasksBackgroundService.syncTasks(
                           ServicesBinding.rootIsolateToken);
                       inject<NotificationBackgroundService>()
-                          .initializeService();
+                          .initializeAndStartService();
                       context.goNamed(Routes.profile.name);
                     }
                   },
