+++ lib/view/pages/feedback/feedback_screen.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:better_one/core/constants/constants.dart';
 import 'package:better_one/core/enum/feedback_cat_enum.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/core/utils/methods/methods.dart';
 import 'package:better_one/core/utils/shared_widgets/back_button_l10n.dart';
