+++ lib/view/pages/onborading/onboarding_screen.dart
@@ -1,6 +1,6 @@
 import 'package:better_one/config/navigation/routes_enum.dart';
 import 'package:better_one/core/constants/constants.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/model/onboarding/onboarding_model.dart';
 import 'package:easy_localization/easy_localization.dart';
