+++ lib/view/pages/task/task_details_screen.dart
@@ -4,7 +4,7 @@ import 'package:better_one/core/constants/comment_constants.dart';
 import 'package:better_one/core/enum/task_status.dart';
 import 'package:better_one/core/in_memory/in_memory.dart';
 import 'package:better_one/core/timer/timer_action.dart';
-import 'package:better_one/core/utils/cache_service/cach_interface/locale_user_info.dart';
+import 'package:better_one/core/utils/service/cache_service/cach_interface/locale_user_info.dart';
 import 'package:better_one/core/utils/dependency_locator/inject.dart';
 import 'package:better_one/core/utils/methods/methods.dart';
 import 'package:better_one/core/utils/shared_widgets/back_button_l10n.dart';
